# Conflicts:
#	nvmolkit/embedMolecules.cpp
#	nvmolkit/embedMolecules.py
#	nvmolkit/mmffOptimization.cpp
#	nvmolkit/mmffOptimization.py
#	nvmolkit/tests/test_mmff_optimization.py
#	nvmolkit/uffOptimization.cpp
#	nvmolkit/uffOptimization.py
#	src/etkdg.cpp
#	src/etkdg.h
#	src/etkdg_impl.h
#	src/etkdg_stage_distgeom_minimize.cu
#	src/etkdg_stage_distgeom_minimize.h
#	src/etkdg_stage_etk_minimization.cu
#	src/etkdg_stage_etk_minimization.h
#	src/minimizer/CMakeLists.txt
#	src/minimizer/bfgs_mmff.h
#	src/minimizer/bfgs_uff.h
#	tests/test_etkdg_etk_minimize.cu
#	tests/test_etkdg_minimize.cu

+++ src/substruct/substruct_search_internal.cpp
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "substruct_search_internal.h"
+#include "src/substruct/substruct_search_internal.h"
 
 #include <GraphMol/ROMol.h>
 #include <GraphMol/Substruct/SubstructMatch.h>
@@ -21,13 +21,13 @@
 #include <mutex>
 #include <vector>
 
-#include "cuda_error_check.h"
-#include "gpu_executor.h"
-#include "nvtx.h"
-#include "pinned_buffer_pool.h"
-#include "substruct_launch_config.h"
-#include "substruct_search.h"
-#include "thread_worker_context.h"
+#include "src/substruct/gpu_executor.h"
+#include "src/substruct/pinned_buffer_pool.h"
+#include "src/substruct/substruct_launch_config.h"
+#include "src/substruct/substruct_search.h"
+#include "src/substruct/thread_worker_context.h"
+#include "src/utils/cuda_error_check.h"
+#include "src/utils/nvtx.h"
 
 namespace nvMolKit {
 
@@ -155,8 +155,8 @@ void processWithRDKitFallback(const RDKit::ROMol*       target,
   if (boolResults) {
     boolResults->setMatch(targetIdx, queryIdx, true);
   } else if (countResults) {
-    const int pairIdx        = targetIdx * results.numQueries + queryIdx;
-    (*countResults)[pairIdx] = matchCount;
+    const int64_t pairIdx                         = static_cast<int64_t>(targetIdx) * results.numQueries + queryIdx;
+    (*countResults)[static_cast<size_t>(pairIdx)] = matchCount;
   } else {
     std::vector<std::vector<int>> convertedMatches;
     convertedMatches.reserve(rdkitMatches.size());
@@ -399,9 +399,9 @@ void accumulateMiniBatchResultsCounts(GpuExecutor&               executor,
   std::lock_guard<std::mutex> lock(resultsMutex);
 
   for (int i = 0; i < executor.plan.numPairsInMiniBatch; ++i) {
-    const auto [targetIdx, queryIdx] = resolvePairIndices(i, ctx, hostBuffer);
-    const int globalPairIdx          = targetIdx * ctx.numQueries + queryIdx;
-    counts[globalPairIdx]            = hostBuffer.matchCounts[i];
+    const auto [targetIdx, queryIdx]           = resolvePairIndices(i, ctx, hostBuffer);
+    const int64_t globalPairIdx                = static_cast<int64_t>(targetIdx) * ctx.numQueries + queryIdx;
+    counts[static_cast<size_t>(globalPairIdx)] = hostBuffer.matchCounts[i];
   }
 }
